+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Exception `QueryIdInvalidException`
 
 
 ## [14.2.0-rc-452] - 2018-02-24